Added Pause Menu on Level 2 and created the Game Over Screen
I have created a functional Pause Menu - accessible with Esc key. The Pause Menu does halt the game - Resume continues from the point it was left off, Restart does the same thing for the moment (until we add Respawn points). Main Menu button and Quit are also functional. The Game Over screen is ready just design-wise, as it will have a script in the moment we add the losing logic.